- removed old smoothing module - renamed menu item for resolvers
