+++ src/main/java/io/github/mzmine/modules/dataprocessing/featdet_smoothing/SmoothingModule.java
@@ -1,88 +1,58 @@
-/*
- * Copyright 2006-2020 The MZmine Development Team
- * 
- * This file is part of MZmine.
- * 
- * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
- * General Public License as published by the Free Software Foundation; either version 2 of the
- * License, or (at your option) any later version.
- * 
- * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License along with MZmine; if not,
- * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
- * USA
- */
-
-/*
- * Code created was by or on behalf of Syngenta and is released under the open source license in use
- * for the pre-existing code or project. Syngenta does not assert ownership or copyright any over
- * pre-existing work.
- */
-
 package io.github.mzmine.modules.dataprocessing.featdet_smoothing;
 
-import io.github.mzmine.datamodel.features.FeatureList;
-import io.github.mzmine.util.MemoryMapStorage;
-import java.util.Collection;
-
-import javax.annotation.Nonnull;
-
 import io.github.mzmine.datamodel.MZmineProject;
+import io.github.mzmine.datamodel.features.ModularFeatureList;
 import io.github.mzmine.modules.MZmineModuleCategory;
 import io.github.mzmine.modules.MZmineProcessingModule;
 import io.github.mzmine.parameters.ParameterSet;
 import io.github.mzmine.taskcontrol.Task;
 import io.github.mzmine.util.ExitCode;
+import io.github.mzmine.util.MemoryMapStorage;
+import java.util.Collection;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 
-/**
- * Chromatographic smoothing.
- * 
- * @version $Revision$
- */
 public class SmoothingModule implements MZmineProcessingModule {
 
-  private static final String MODULE_NAME = "Smoothing";
-  private static final String MODULE_DESCRIPTION =
-      "This module performs smoothing of chromatograms prior to deconvolution.";
+  private static final String name = "Smoothing";
 
-  @Override
-  public @Nonnull String getName() {
-    return MODULE_NAME;
-  }
-
-  @Override
-  public @Nonnull String getDescription() {
-    return MODULE_DESCRIPTION;
-  }
-
-  @Override
   @Nonnull
+  @Override
+  public String getName() {
+    return name;
+  }
+
+  @Nullable
+  @Override
+  public Class<? extends ParameterSet> getParameterSetClass() {
+    return SmoothingParameters.class;
+  }
+
+  @Nonnull
+  @Override
+  public String getDescription() {
+    return "Smoothes intensity along the retentio time and/or mobility dimension.";
+  }
+
+  @Nonnull
+  @Override
   public ExitCode runModule(@Nonnull MZmineProject project, @Nonnull ParameterSet parameters,
       @Nonnull Collection<Task> tasks) {
-    FeatureList peakLists[] =
-        parameters.getParameter(SmoothingParameters.peakLists).getValue().getMatchingFeatureLists();
+
+    final ModularFeatureList[] flists = parameters.getParameter(SmoothingParameters.featureLists)
+        .getValue().getMatchingFeatureLists();
 
     final MemoryMapStorage storage = MemoryMapStorage.forFeatureList();
-
-    for (final FeatureList peakList : peakLists) {
-      Task newTask = new SmoothingTask(project, peakList, parameters, storage);
-      tasks.add(newTask);
+    for (ModularFeatureList flist : flists) {
+      tasks.add(new SmoothingTask(project, flist, storage, parameters));
     }
 
     return ExitCode.OK;
   }
 
+  @Nonnull
   @Override
-  public @Nonnull MZmineModuleCategory getModuleCategory() {
-    return MZmineModuleCategory.FEATURELISTDETECTION;
+  public MZmineModuleCategory getModuleCategory() {
+    return MZmineModuleCategory.FEATUREDETECTION;
   }
-
-  @Override
-  public @Nonnull Class<? extends ParameterSet> getParameterSetClass() {
-    return SmoothingParameters.class;
-  }
-
 }
