--- src/main/java/io/github/mzmine/modules/dataprocessing/featdet_smoothing/help/help.html
@@ -1,39 +1,0 @@
-<!--
-  Code created was by or on behalf of Syngenta and is released under the open source license in use for the
-  pre-existing code or project. Syngenta does not assert ownership or copyright any over pre-existing work.
-  -->
-<html>
-<head>
-    <title>Peak detection - Smoothing</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
-    <link rel="stylesheet" type="text/css" href="/net/sf/mzmine/desktop/impl/helpsystem/HelpStyles.css">
- </head>
-
-<body>
-
-<h1>Smoothing</h1>
-
-<h2>Description</h2>
-
-<p>
-    This module applies Savitzky-Golay smoothing to the peak chromatograms in a peak list. Its main purpose is to
-    remove high frequency noise from the chromatograms prior to deconvolution.
-</p>
-
-<h4>Method parameters</h4>
-
-<dl>
-    <dt>Filename suffix</dt>
-    <dd>The text to append to the name of the smoothed peak list.</dd>
-
-    <dt>Filter width</dt>
-    <dd>The number of data points over which the filter kernel extends. The wider the filter the greater the smoothing
-        effect.
-    </dd>
-
-    <dt>Remove original peak list</dt>
-    <dd>Whether to remove the original peak list once smoothing is complete.</dd>
-</dl>
-
-</body>
-</html>
